api/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
flupy==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
gotrue==2.6.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
llama-cloud==0.0.13
llama-index==0.10.64
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.64
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.29
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-supabase==0.15
llama-parse==0.4.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.2
numpy==1.26.4
openai==1.40.2
packaging==24.1
pandas==2.2.2
pgvector==0.1.8
pillow==10.4.0
postgrest==0.16.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
realtime==1.0.6
regex==2024.7.24
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
striprtf==0.0.26
supabase==2.6.0
supafunc==0.5.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tqdm==4.66.5
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0; sys_platform != 'win32'
vecs==0.4.4
watchfiles==0.23.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
pytest~=8.3.2